backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
allosaurus==1.0.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
csvw==3.5.1
distro==1.9.0
dlinfo==2.0.0
editdistance==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language-tags==1.2.0
llvmlite==0.44.0
MarkupSafe==3.0.2
mistralai==0.4.2
mpmath==1.3.0
multidict==6.1.0
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.66.3
orjson==3.10.15
panphon==0.21.2
phonemizer==3.3.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
resampy==0.4.3
rfc3986==1.5.0
rpds-py==0.23.1
scipy==1.15.2
segments==2.3.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.3
yarl==1.18.3
